+++ data/info.txt
@@ -1,6 +1,4 @@
 
-Name: Jeremy Phone #: 6161626
-Name: Bethany Phone #: 832765
-Name: Jeff Phone #: 6356
-Name: ge Phone #: fer
-Name: jEFF Phone #: 43532
+Name: Jim    Phone #: (215) 253-5234
+Name: James    Phone #: (678) 765-4345
+Name: James Lampkins    Phone #: (123) 456-7891
